+++ Demo/Application/Base/ContainmentViewController.swift
@@ -176,6 +176,11 @@ class ContainmentViewController: UIViewController {
                 }
             }
 
+        case .newPayPalCheckoutTokenizationKey:
+            updateStatus("Fetching new checkout token...")
+            let newPayPalCheckoutTokenizationKey = "sandbox_rz48bqvw_jcyycfw6f9j4nj9c"
+            currentViewController = instantiateViewController(with: newPayPalCheckoutTokenizationKey)
+
         case .mockedPayPalTokenizationKey:
             let tokenizationKey = "sandbox_q7v35n9n_555d2htrfsnnmfb3"
             currentViewController = instantiateViewController(with: tokenizationKey)
@@ -208,6 +213,8 @@ class ContainmentViewController: UIViewController {
             return PayPalWebCheckoutViewController(authorization: authorization)
         case "SEPADirectDebitViewController":
             return SEPADirectDebitViewController(authorization: authorization)
+        case "ShopperInsightsViewController":
+            return ShopperInsightsViewController(authorization: authorization)
         case "ThreeDSecureViewController":
             return ThreeDSecureViewController(authorization: authorization)
         case "VenmoViewController":