+++ Sources/BraintreeCore/BTAPIClient.swift
@@ -20,6 +20,7 @@ import Foundation
     
     var http: BTHTTP?
     var graphQLHTTP: BTGraphQLHTTP?
+    var payPalHTTP: BTHTTP?
 
     /// Exposed for testing analytics
     /// By default, the `BTAnalyticsService` instance is static/shared so that only one queue of events exists.
@@ -285,6 +286,7 @@ import Foundation
     public func post(
         _ path: String,
         parameters: Encodable,
+        headers: [String: String]? = nil,
         httpType: BTAPIClientHTTPService = .gateway,
         completion: @escaping RequestCompletion
     ) {
@@ -300,7 +302,33 @@ import Foundation
             }
 
             let postParameters = BTAPIRequest(requestBody: parameters, metadata: metadata, httpType: httpType)
-            http(for: httpType)?.post(path, configuration: configuration, parameters: postParameters, completion: completion)
+            http(for: httpType)?.post(path, configuration: configuration, parameters: postParameters, headers: headers, completion: completion)
+        }
+    }
+    
+    /// :nodoc: This method is exposed for internal Braintree use only. Do not use. It is not covered by Semantic Versioning and may change or be removed at any time.
+    ///
+    /// Perfom an HTTP POST on a URL composed of the configured from environment and the given path.
+    /// - Parameters:
+    ///   - path: The endpoint URI path.
+    ///   - parameters: Optional set of query parameters to be encoded with the request.
+    ///   - httpType: The underlying `BTAPIClientHTTPService` of the HTTP request. Defaults to `.gateway`.
+    /// - Returns: On success, `(BTJSON?, HTTPURLResponse?)` will contain the JSON body response and the HTTP response.
+    @_documentation(visibility: private)
+    public func post(
+        _ path: String,
+        parameters: Encodable,
+        headers: [String: String]? = nil,
+        httpType: BTAPIClientHTTPService = .gateway
+    ) async throws -> (BTJSON?, HTTPURLResponse?) {
+        try await withCheckedThrowingContinuation { continuation in
+            post(path, parameters: parameters, headers: headers, httpType: httpType) { json, httpResonse, error in
+                if let error {
+                    continuation.resume(throwing: error)
+                } else {
+                    continuation.resume(returning: (json, httpResonse))
+                }
+            }
         }
     }
 
@@ -333,6 +361,8 @@ import Foundation
             return parameters?.merging(["_meta": metadata.parameters]) { $1 }
         case .graphQLAPI:
             return parameters?.merging(["clientSdkMetadata": metadata.parameters]) { $1 }
+        case .payPalAPI:
+            return parameters
         }
     }
 
@@ -355,6 +385,8 @@ import Foundation
             return http
         case .graphQLAPI:
             return graphQLHTTP
+        case .payPalAPI:
+            return payPalHTTP
         }
     }
     
@@ -369,6 +401,24 @@ import Foundation
         if graphQLHTTP == nil {
             graphQLHTTP = BTGraphQLHTTP(authorization: authorization)
             graphQLHTTP?.networkTimingDelegate = self
+        }
+        
+        if payPalHTTP == nil {
+            let paypalBaseURL: URL? = payPalAPIURL(forEnvironment: configuration.environment ?? "")
+            
+            if authorization.type == .clientToken {
+                payPalHTTP = BTHTTP(authorization: authorization, customBaseURL: paypalBaseURL)
+                payPalHTTP?.networkTimingDelegate = self
+            }
+        }
+    }
+    
+    func payPalAPIURL(forEnvironment environment: String) -> URL? {
+        if environment.caseInsensitiveCompare("sandbox") == .orderedSame ||
+            environment.caseInsensitiveCompare("development") == .orderedSame {
+            return BTCoreConstants.payPalSandboxURL
+        } else {
+            return BTCoreConstants.payPalProductionURL
         }
     }