+++ Sources/BraintreePayPal/BTPayPalCheckoutRequest.swift
@@ -81,6 +81,9 @@ import BraintreeCore
 
     /// Optional: If set to `true`, this enables the Checkout with Vault flow, where the customer will be prompted to consent to a billing agreement during checkout. Defaults to `false`.
     public var requestBillingAgreement: Bool
+    
+    /// Optional: User email to initiate a quicker authentication flow in cases where the user has a PayPal Account with the same email.
+    public var userAuthenticationEmail: String?
 
     // MARK: - Initializer
 
@@ -129,6 +132,10 @@ import BraintreeCore
 
         if currencyCode != nil {
             checkoutParameters["currency_iso_code"] = currencyCode
+        }
+        
+        if let userAuthenticationEmail, !userAuthenticationEmail.isEmpty {
+            checkoutParameters["payer_email"] = userAuthenticationEmail
         }
 
         if userAction != .none, var experienceProfile = baseParameters["experience_profile"] as? [String: Any] {