+++ UnitTests/BraintreeTestShared/FakeHTTP.swift
@@ -6,6 +6,7 @@ import Foundation
     @objc public var POSTRequestCount: Int = 0
     @objc public var lastRequestEndpoint: String?
     public var lastRequestMethod: String?
+    public var lastPOSTRequestHeaders: [String: String]? = [:]
     @objc public var lastRequestParameters: [String: Any]?
     var stubMethod: String?
     var stubEndpoint: String?
@@ -54,11 +55,12 @@ import Foundation
         }
     }
 
-    public override func post(_ path: String, configuration: BTConfiguration? = nil, parameters: [String: Any]? = nil, completion: ((BTJSON?, HTTPURLResponse?, Error?) -> Void)? = nil) {
+    public override func post(_ path: String, configuration: BTConfiguration? = nil, parameters: [String: Any]? = nil, headers: [String: String]? = nil, completion: ((BTJSON?, HTTPURLResponse?, Error?) -> Void)? = nil) {
         POSTRequestCount += 1
         lastRequestEndpoint = path
         lastRequestParameters = parameters
         lastRequestMethod = "POST"
+        lastPOSTRequestHeaders = headers
         if cannedError != nil {
             dispatchQueue.async {
                 completion?(nil, nil, self.cannedError)
@@ -82,7 +84,7 @@ import Foundation
         return self.init(authorization: fakeTokenizationKey, customBaseURL: URL(string: "http://fake.com")!)
     }
 
-    public override func post(_ path: String, configuration: BTConfiguration? = nil, parameters: [String: Any]?, completion: ((BTJSON?, HTTPURLResponse?, Error?) -> Void)? = nil) {
+    public override func post(_ path: String, configuration: BTConfiguration? = nil, parameters: [String: Any]?, headers: [String: String]? = nil, completion: ((BTJSON?, HTTPURLResponse?, Error?) -> Void)? = nil) {
         POSTRequestCount += 1
         lastRequestParameters = parameters
         completion?(self.cannedConfiguration, nil, nil)